+++ dask_saturn/__init__.py
@@ -4,6 +4,7 @@ imports added so users do not have to think about submodules
 
 from .core import describe_sizes, list_sizes, SaturnCluster  # noqa: F401
 from ._version import get_versions
+from .plugins import SaturnSetup
 
 __version__ = get_versions()["version"]
 del get_versions
